chg: dev: rename module, refactor package metadata, move to src layout
