+++ pyproject.toml
@@ -5,35 +5,98 @@ requires = [
 ]
 build-backend = "setuptools.build_meta"
 
+[project]
+name = "bmaptool"
+description = "BMAP tools"
+dynamic = ["version"]
+dependencies = [
+    "defusedxml",
+    "gpg >= 1.10.0",
+    'importlib-metadata>=4.6; python_version < "3.10"',
+]
+requires-python = ">=3.8"
+authors = [
+    {name = "Joshua Watt", email = "JPEWhacker@gmail.com"},
+    {name = "Trevor Woerner", email = "twoerner@gmail.com"},
+    {name = "Tim Orling", email = "ticotimo@gmail.com"},
+
+]
+readme = "README.md"
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Developers",
+    "Topic :: Software Development :: Build Tools",
+    "Topic :: Software Development :: Embedded Systems",
+    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+]
+
+[project.optional-dependencies]
+dev = [
+    "black >= 22.3.0",
+    "six >= 1.16.0",
+]
+cov = [
+    "coverage",
+    "coverage_python_version",
+]
+doc = [
+    "sphinx",
+    "sphinx_git",
+    "recommonmark",
+    "sphinx_rtd_theme",
+    "sphinxcontrib-apidoc",
+]
+test = [
+    "pytest",
+    "pytest-cov",
+]
+
+[project.urls]
+Homepage = "https://github.com/yoctoproject/bmaptool"
+Repository = "https://github.com/yoctoproject/bmaptool.git"
+Issues = "https://github.com/yoctoproject/bmaptool/issues"
+Changelog = "https://github.com/yoctoproject/bmaptool/blob/master/CHANGELOG.md"
+
+[project.scripts]
+bmaptool = "bmaptool.CLI:main"
+
 [tool.setuptools_scm]
 
 [tool.pytest.ini_options]
 minversion = "6.0"
-testpaths = ["tests",]
-log_cli =  true
+testpaths = ["tests", "src/bmaptool",]
+log_cli =  false
 doctest_optionflags = ["ELLIPSIS", "NORMALIZE_WHITESPACE",]
-addopts = "--strict-markers"
+addopts = "-ra -q -p no:warnings"
 markers = "subscript"
 
 [tool.coverage.run]
 branch = true
-include = ["bmaptools/"]
-source = [
-    "bmaptools",
-    ".tox/py*/lib/python*/site-packages/",
-]
-omit = [
-    "docs",
-    "tests",
-    ".tox",
-]
+source = ["src/bmaptool"]
+plugins = ["coverage_python_version"]
 
 [tool.coverage.paths]
-source = ["bmaptools"]
+source = [
+    "src/",
+    "*/site-packages"
+]
 
 [tool.coverage.report]
 fail_under = 40
 show_missing = true
+exclude_lines = [
+    "pragma: no cover",
+    "raise NotImplementedError",
+    "raise AssertionError",
+    "if typing.TYPE_CHECKING:",
+    "if TYPE_CHECKING:",
+]
 
 [tool.black]
 line-length = 90
