--- setup.cfg
@@ -1,80 +1,0 @@
-[metadata]
-name = bmaptools
-version = attr: setuptools_scm.get_version
-description = Tools to generate block map (AKA bmap) and copy images using bmap
-url = https://github.com/intel/bmap-tools
-author = Artem Bityutskiy
-author_email = artem.bityutskiy@linux.intel.com
-maintainer = Simon McVittie
-maintainer_email = smcv@debian.org
-long_description = file: README.md
-long_description_content_type = text/markdown; charset=UTF-8
-license_expression = GPL-2.0-only
-license_files = LICENSE
-classifiers =
-    Intended Audience :: Developers
-    Programming Language :: Python
-    Programming Language :: Python :: 3
-    Environment :: Console
-    Topic :: Software Development
-
-[options]
-python_requires = >=3.7
-
-install_requires =
-    defusedxml
-    gpg
-    importlib-metadata; python_version < '3.8'
-    six
-
-packages = bmaptools
-
-[options.entry_points]
-console_scripts =
-    bmaptool = bmaptools.CLI:main
-
-# deps are included here mainly for local/venv installs using pip
-# otherwise deps are handled via tox, ci config files or pkg managers
-[options.extras_require]
-doc =
-    sphinx
-    sphinx_git
-    recommonmark
-    sphinx_rtd_theme
-    sphinxcontrib-apidoc
-test =
-    pytest
-    pytest-cov
-cov =
-    coverage[toml]
-    coverage_python_version
-all =
-    %(cov)s
-    %(doc)s
-    %(test)s
-
-[check]
-metadata = true
-restructuredtext = true
-strict = false
-
-[check-manifest]
-ignore =
-    .gitattributes
-    .gitignore
-    .pre-commit-config.yaml
-
-[flake8]
-exclude =
-    .git,
-    __pycache__,
-    build,
-    contrib,
-    debian,
-    dist,
-    docs,
-    packaging,
-    tests
-
-max-line-length = 90
-#extend-ignore = E203
