+++ src/bmaptool/BmapCopy.py
