+++ src/bmaptool/BmapCreate.py
@@ -48,8 +48,8 @@ This module uses the FIEMAP ioctl to detect holes.
 # pylint: disable=R0902,R0903
 
 import hashlib
-from bmaptools.BmapHelpers import human_size
-from bmaptools import Filemap
+from bmaptool.BmapHelpers import human_size
+from bmaptool import Filemap
 
 # The bmap format version we generate.
 #
