+++ src/bmaptool/BmapHelpers.py
