+++ tox.ini
@@ -1,21 +1,19 @@
 [tox]
-envlist = py3{6,7,8,9,10,11}-{linux,macos,windows}
+envlist = py3{9,10,11,12}-{linux,macos,windows}
 skip_missing_interpreters = true
 isolated_build = true
 skipsdist = true
 
 [gh-actions]
 python =
-    3.6: py36
-    3.7: py37
-    3.8: py38
     3.9: py39
     3.10: py310
     3.11: py311
+    3.12: py312
 
 [gh-actions:env]
 PLATFORM =
-    ubuntu-20.04: linux
+    ubuntu-22.04: linux
     macos-latest: macos
     windows-latest: windows
 
@@ -35,6 +33,7 @@ skip_install = true
 
 setenv =
     COVERAGE_FILE = .coverage.{envname}
+    PYTHONPATH = {toxinidir}/src
 
 passenv =
     HOME
@@ -51,11 +50,10 @@ allowlist_externals =
 
 deps =
     {[base]deps}
-    #-r requirements.txt
-    .[test,cov]
+    .[test,cov,dev]
 
 commands =
-    python -m pytest -v --basetemp="{envtmpdir}" --cov=bmaptools --cov-report term-missing {posargs}
+    python -m pytest -v . --capture=fd --cov=bmaptool --cov-branch --cov-report term-missing {posargs}
 
 [testenv:coverage]
 basepython =
@@ -75,12 +73,10 @@ commands =
     coverage xml
 
 depends =
-    py36
-    py37
-    py38
     py39
     py310
     py311
+    py312
 
 [coverage:run]
 parallel=True
@@ -107,7 +103,7 @@ allowlist_externals =
 deps =
     {[base]deps}
     -r requirements.txt
-    -e .
+    -e .[dev]
 
 commands =
     pytest -v --basetemp="{envtmpdir}" {posargs}
@@ -124,7 +120,7 @@ passenv =
 
 deps =
     pip>=21.1
-    https://github.com/sarnold/gitchangelog/releases/download/3.1.2/gitchangelog-3.1.2-py3-none-any.whl
+    https://github.com/sarnold/gitchangelog/releases/download/3.2.0/gitchangelog-3.2.0-py3-none-any.whl
 
 commands =
     bash -c 'gitchangelog {posargs} > CHANGELOG.md'
@@ -139,7 +135,7 @@ deps =
 
 commands =
     #bash -c 'make -C docs/ clean'
-    bash -c 'rm -rf *.egg-info dump.* .coverage.* bmaptools/__pycache__ dist/ build/ docs/source/api/'
+    bash -c 'rm -rf src/*.egg-info dump.* .coverage.* src/*/__pycache__ dist/ build/ docs/source/api/'
 
 [testenv:docs]
 skip_install = true
@@ -189,7 +185,7 @@ deps =
     pip>=21.1
 
 commands =
-    pip install bmaptools --force-reinstall --pre --prefer-binary -f dist/
+    pip install bmaptool --force-reinstall --pre --prefer-binary -f dist/
     bmaptool --version
 
 [testenv:lint]
@@ -204,7 +200,7 @@ deps =
     -r requirements.txt
 
 commands =
-    pylint --fail-under=8.0 --ignore=_version.py bmaptools/
+    pylint --fail-under=8.0 --ignore=_version.py src/bmaptools/
 
 [testenv:style]
 passenv =
@@ -218,7 +214,7 @@ deps =
     flake8-bugbear
 
 commands =
-    flake8 bmaptools/
+    flake8 src/bmaptools/
 
 [testenv:mypy]
 skip_install = true
@@ -231,7 +227,7 @@ deps =
     -r requirements.txt
 
 commands =
-    python -m mypy --follow-imports=normal --install-types --non-interactive bmaptools/
+    python -m mypy --follow-imports=normal --install-types --non-interactive src/bmaptools/
 
 [testenv:black]
 skip_install = true
@@ -243,7 +239,7 @@ deps =
     black
 
 commands =
-    black -v -S bmaptools/ {posargs}
+    black -v -S src/bmaptools/ {posargs}
 
 [testenv:sec]
 skip_install = true
@@ -259,7 +255,7 @@ deps =
     bandit[toml]
 
 commands =
-    bandit -c pyproject.toml -r bmaptools/
+    bandit -c pyproject.toml -r src/bmaptools/
 
 [testenv:isort]
 skip_install = true
@@ -272,4 +268,4 @@ deps =
     -r requirements.txt
 
 commands =
-    python -m isort bmaptools/
+    python -m isort src/bmaptools/
